Restore Cleanup function
