+++ robottelo/cleanup.py
@@ -6,7 +6,7 @@ from nailgun import entities
 from robottelo.cli.base import CLIReturnCodeError
 from robottelo.cli.proxy import Proxy
 from robottelo.decorators import bz_bug_is_open
-
+from robottelo.vm import VirtualMachine
 
 LOGGER = logging.getLogger(__name__)
 
@@ -57,3 +57,22 @@ def vm_cleanup(vm):
     :param robottelo.vm.VirtualMachine vm: virtual machine to destroy
     """
     vm.destroy()
+
+
+def cleanup_of_provisioned_server(hostname=None, provisioning_server=None,
+                                  distro=None):
+    """ Cleanup the VM from provisioning server
+
+    :param: str hostname: The content host hostname
+    :param: str provisioning_server: provision server name
+    :param: str distro: distro type
+    """
+    if hostname:
+        vm = VirtualMachine(
+            hostname=hostname,
+            target_image=hostname,
+            provisioning_server=provisioning_server,
+            distro=distro,
+        )
+        vm._created = True
+        vm.destroy()
