+++ tests/upgrades/test_remoteexecution.py
@@ -15,9 +15,9 @@
 :Upstream: No
 """
 from nailgun import entities
+from robottelo.cleanup import cleanup_of_provisioned_server
 from robottelo.constants import DEFAULT_LOC, DEFAULT_ORG, DISTRO_RHEL7
 from robottelo.helpers import add_remote_execution_ssh_key
-from robottelo.upgrade_utility import CommonUpgradeUtility
 from robottelo.vm import VirtualMachine
 from robottelo.test import APITestCase, settings
 from upgrade_tests import post_upgrade, pre_upgrade
@@ -112,7 +112,7 @@ class Scenario_remoteexecution_external_capsule(APITestCase):
             create_dict(global_dict)
         except Exception as exp:
             if client._created:
-                CommonUpgradeUtility.cleanup_of_provisioned_server(
+                cleanup_of_provisioned_server(
                     hostname=client.hostname,
                     provisioning_server=self.libvirt_vm,
                     distro=DISTRO_RHEL7)
@@ -136,7 +136,7 @@ class Scenario_remoteexecution_external_capsule(APITestCase):
             'job_template_id': 89, 'inputs': {'command': "ls"},
             'targeting_type': 'static_query', 'search_query': "name = {0}".format(client_name)})
         self.assertEqual(job['output']['success_count'], 1)
-        CommonUpgradeUtility.cleanup_of_provisioned_server(
+        cleanup_of_provisioned_server(
             hostname=client_name,
             provisioning_server=self.libvirt_vm,
             distro=DISTRO_RHEL7)
@@ -228,7 +228,7 @@ class Scenario_remoteexecution_satellite(APITestCase):
             create_dict(global_dict)
         except Exception as exp:
             if client._created:
-                CommonUpgradeUtility.cleanup_of_provisioned_server(
+                cleanup_of_provisioned_server(
                     hostname=client.hostname,
                     provisioning_server=self.libvirt_vm,
                     distro=DISTRO_RHEL7)
@@ -253,7 +253,7 @@ class Scenario_remoteexecution_satellite(APITestCase):
             'targeting_type': 'static_query', 'search_query': "name = {0}".
             format(client_name)})
         self.assertEqual(job['output']['success_count'], 1)
-        CommonUpgradeUtility.cleanup_of_provisioned_server(
+        cleanup_of_provisioned_server(
             hostname=client_name,
             provisioning_server=self.libvirt_vm,
             distro=DISTRO_RHEL7)
